added more fo loadmap.js
Added some pseudo code to the beiginning to make sure we are adding the features or attempting to add some of the features we metioned in the project proposal. I'm starting to think we made need a side bar that controls all the features. Also, added some code that will be the pop-windows

added Jquery to the jsp file.